Material应用，拆分widget
