+++ lib/demo/hello.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+class Hello extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+          child: Text(
+            'Hello Flutter', 
+            textDirection: TextDirection.ltr,
+            style: TextStyle(
+              fontSize: 40,
+              fontWeight: FontWeight.bold,
+              color: Colors.black45
+            )
+            ),
+        );
+  }
+}
