+++ lib/demo/listviewitem.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import '../model/post.dart';
+class ListViewItem extends StatelessWidget {
+  Widget _listItemBuilder(BuildContext context, int index) {
+    return Container(
+      color: Colors.white10,
+      margin: EdgeInsets.all(8.0),
+      child: Column(
+        children: <Widget>[
+          Image.network(posts[index].imageUrl),
+          SizedBox(height: 16.0),
+          Text(
+            posts[index].title,
+            style: Theme.of(context).textTheme.title,
+          ),
+          Text(
+            posts[index].author,
+            style: Theme.of(context).textTheme.subhead,
+          ),
+          SizedBox(height: 16.0)
+        ],
+      ),
+    );
+  }
+  @override
+  Widget build(BuildContext context) {
+    return ListView.builder(
+          itemCount: posts.length,
+          itemBuilder: _listItemBuilder,
+        );
+  }
+}
