+++ lib/main.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
-import 'model/post.dart';
+// import './demo/listviewitem.dart';
 void main() => runApp(MyApp());
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       home: Home(),
       theme: ThemeData(
         primarySwatch: Colors.yellow
@@ -13,27 +14,6 @@ class MyApp extends StatelessWidget {
   }
 }
 class Home extends StatelessWidget {
-  Widget _listItemBuilder(BuildContext context, int index) {
-    return Container(
-      color: Colors.white10,
-      margin: EdgeInsets.all(8.0),
-      child: Column(
-        children: <Widget>[
-          Image.network(posts[index].imageUrl),
-          SizedBox(height: 16.0),
-          Text(
-            posts[index].title,
-            style: Theme.of(context).textTheme.title,
-          ),
-          Text(
-            posts[index].author,
-            style: Theme.of(context).textTheme.subhead,
-          ),
-          SizedBox(height: 16.0)
-        ],
-      ),
-    );
-  }
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -42,26 +22,7 @@ class Home extends StatelessWidget {
           title: Text('MICHAEL CHEUNG'),
           elevation: 1.0,
         ),
-        body: ListView.builder(
-          itemCount: posts.length,
-          itemBuilder: _listItemBuilder,
-        ),
+        body: null
       );
-  }
-}
-class Hello extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-          child: Text(
-            'Hello Flutter', 
-            textDirection: TextDirection.ltr,
-            style: TextStyle(
-              fontSize: 40,
-              fontWeight: FontWeight.bold,
-              color: Colors.black45
-            )
-            ),
-        );
   }
 }
